requirements.txt
allure-pytest
requests
pytest
pathlib